+++ src/api/my_account.rs
@@ -87,7 +87,11 @@ mod test {
     #[test]
     fn test_get_my_account() -> Result<(), Box<dyn Error>> {
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = GetMyAccount::builder().build()?;
         redmine.json_response_body::<_, UserWrapper<MyAccount>>(&endpoint)?;
         Ok(())
@@ -101,7 +105,11 @@ mod test {
     #[test]
     fn test_completeness_my_account_type() -> Result<(), Box<dyn Error>> {
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = GetMyAccount::builder().build()?;
         let UserWrapper { user: value } =
             redmine.json_response_body::<_, UserWrapper<serde_json::Value>>(&endpoint)?;