+++ src/api/news.rs
@@ -119,7 +119,11 @@ mod test {
     #[test]
     fn test_list_news_first_page() -> Result<(), Box<dyn Error>> {
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = ListNews::builder().build()?;
         redmine.json_response_body_page::<_, News>(&endpoint, 0, 25)?;
         Ok(())
@@ -129,7 +133,11 @@ mod test {
     #[test]
     fn test_list_news_all_pages() -> Result<(), Box<dyn Error>> {
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = ListNews::builder().build()?;
         redmine.json_response_body_all_pages::<_, News>(&endpoint)?;
         Ok(())
@@ -143,7 +151,11 @@ mod test {
     #[test]
     fn test_completeness_news_type() -> Result<(), Box<dyn Error>> {
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = ListNews::builder().build()?;
         let values: Vec<serde_json::Value> = redmine.json_response_body_all_pages(&endpoint)?;
         for value in values {