+++ src/api/projects.rs
@@ -556,7 +556,11 @@ pub(crate) mod test {
     fn test_list_projects_first_page() -> Result<(), Box<dyn Error>> {
         let _r_project = PROJECT_LOCK.read();
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = ListProjects::builder().build()?;
         redmine.json_response_body_page::<_, Project>(&endpoint, 0, 25)?;
         Ok(())
@@ -567,7 +571,11 @@ pub(crate) mod test {
     fn test_list_projects_all_pages() -> Result<(), Box<dyn Error>> {
         let _r_project = PROJECT_LOCK.read();
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = ListProjects::builder().build()?;
         redmine.json_response_body_all_pages::<_, Project>(&endpoint)?;
         Ok(())
@@ -578,7 +586,9 @@ pub(crate) mod test {
     async fn test_list_projects_async_first_page() -> Result<(), Box<dyn Error>> {
         let _r_project = PROJECT_LOCK.read();
         dotenvy::dotenv()?;
-        let redmine = crate::api::RedmineAsync::from_env()?;
+        let redmine = crate::api::RedmineAsync::from_env(
+            reqwest::Client::builder().use_rustls_tls().build()?,
+        )?;
         let endpoint = ListProjects::builder().build()?;
         redmine
             .json_response_body_page::<_, Project>(&endpoint, 0, 25)
@@ -591,7 +601,9 @@ pub(crate) mod test {
     async fn test_list_projects_async_all_pages() -> Result<(), Box<dyn Error>> {
         let _r_project = PROJECT_LOCK.read();
         dotenvy::dotenv()?;
-        let redmine = crate::api::RedmineAsync::from_env()?;
+        let redmine = crate::api::RedmineAsync::from_env(
+            reqwest::Client::builder().use_rustls_tls().build()?,
+        )?;
         let endpoint = ListProjects::builder().build()?;
         redmine
             .json_response_body_all_pages::<_, Project>(&endpoint)
@@ -604,7 +616,11 @@ pub(crate) mod test {
     fn test_get_project() -> Result<(), Box<dyn Error>> {
         let _r_project = PROJECT_LOCK.read();
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = GetProject::builder()
             .project_id_or_name("sandbox")
             .build()?;
@@ -646,7 +662,11 @@ pub(crate) mod test {
     fn test_completeness_project_type() -> Result<(), Box<dyn Error>> {
         let _r_project = PROJECT_LOCK.read();
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = ListProjects::builder().build()?;
         let values: Vec<serde_json::Value> = redmine.json_response_body_all_pages(&endpoint)?;
         for value in values {
@@ -672,7 +692,11 @@ pub(crate) mod test {
     {
         let _r_project = PROJECT_LOCK.read();
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = ListProjects::builder()
             .include(vec![
                 ProjectsInclude::Trackers,