+++ src/api/queries.rs
@@ -75,7 +75,11 @@ mod test {
     #[test]
     fn test_list_queries_first_page() -> Result<(), Box<dyn Error>> {
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = ListQueries::builder().build()?;
         redmine.json_response_body_page::<_, Query>(&endpoint, 0, 25)?;
         Ok(())
@@ -85,7 +89,11 @@ mod test {
     #[test]
     fn test_list_queries_all_pages() -> Result<(), Box<dyn Error>> {
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = ListQueries::builder().build()?;
         redmine.json_response_body_all_pages::<_, Query>(&endpoint)?;
         Ok(())
@@ -99,7 +107,11 @@ mod test {
     #[test]
     fn test_completeness_query_type() -> Result<(), Box<dyn Error>> {
         dotenvy::dotenv()?;
-        let redmine = crate::api::Redmine::from_env()?;
+        let redmine = crate::api::Redmine::from_env(
+            reqwest::blocking::Client::builder()
+                .use_rustls_tls()
+                .build()?,
+        )?;
         let endpoint = ListQueries::builder().build()?;
         let values: Vec<serde_json::Value> = redmine.json_response_body_all_pages(&endpoint)?;
         for value in values {